+++ demo/src/js/demo.js
@@ -279,24 +279,12 @@ import Raven from 'raven-js';
 
     // Google analytics
     // For demo site (https://plyr.io) only
-    /* eslint-disable */
     if (env.prod) {
-        ((i, s, o, g, r, a, m) => {
-            i.GoogleAnalyticsObject = r;
-            i[r] =
-                i[r] ||
-                function() {
-                    (i[r].q = i[r].q || []).push(arguments);
-                };
-            i[r].l = 1 * new Date();
-            a = s.createElement(o);
-            m = s.getElementsByTagName(o)[0];
-            a.async = 1;
-            a.src = g;
-            m.parentNode.insertBefore(a, m);
-        })(window, document, 'script', 'https://www.google-analytics.com/analytics.js', 'ga');
-        window.ga('create', 'UA-40881672-11', 'auto');
-        window.ga('send', 'pageview');
+        window.dataLayer = window.dataLayer || [];
+        const gtag = (...args) => {
+            window.dataLayer.push(args);
+        };
+        gtag('js', new Date());
+        gtag('config', 'UA-132699580-1');
     }
-    /* eslint-enable */
 })();